Simplify packer by separating packable and pack logic as well as excel operations out
