+++ tests/models/test_scenario_packer.py
@@ -5,7 +5,6 @@ import os
 from unittest.mock import Mock, patch
 from pyetm.models.scenario_packer import (
     ScenarioPacker,
-    ExportConfigResolver,
 )
 from pyetm.models.packables.custom_curves_pack import CustomCurvesPack
 from pyetm.models.packables.inputs_pack import InputsPack
@@ -13,8 +12,8 @@ from pyetm.models.packables.output_curves_pack import OutputCurvesPack
 from pyetm.models.packables.sortable_pack import SortablePack
 from pyetm.models.packables.query_pack import QueryPack
 from pyetm.models import Scenario
-from pyetm.models.custom_curves import CustomCurves
 from pyetm.models.export_config import ExportConfig
+from pyetm.utils.excel_utils import ExportConfigResolver
 
 
 class TestScenarioPackerInit:
@@ -628,58 +627,6 @@ class TestExportConfigResolver:
 
 class TestScenarioPackerHelpers:
 
-    def test_find_first_non_empty_row(self):
-        """Test _find_first_non_empty_row method"""
-        packer = ScenarioPacker()
-
-        assert packer._find_first_non_empty_row(None) is None
-
-        empty = pd.DataFrame([[float("nan")], [float("nan")]])
-        assert packer._find_first_non_empty_row(empty) is None
-
-        # Test with actual data
-        df = pd.DataFrame([[None, None], ["header", "value"], [1, 2]])
-        assert packer._find_first_non_empty_row(df) == 1
-
-    def test_is_helper_column(self):
-        """Test _is_helper_column method"""
-        packer = ScenarioPacker()
-        helpers = {"sortables", "hour", "index"}
-
-        assert packer._is_helper_column(123, helpers) is True
-        assert packer._is_helper_column(" ", helpers) is True
-        assert packer._is_helper_column("NaN", helpers) is True
-        assert packer._is_helper_column("hour", helpers) is True
-        assert packer._is_helper_column("value", helpers) is False
-
-    def test_normalize_sheet(self):
-        """Test _normalize_sheet method"""
-        packer = ScenarioPacker()
-
-        # None -> empty
-        assert packer._normalize_sheet(None, helper_names=set()).empty
-
-        # Build a frame with header at row 1 (0-based)
-        raw = pd.DataFrame(
-            [
-                [None, None, None],
-                ["index", "heat_network", "value"],  # header
-                [1, "hn", 10],
-                [2, "hn", 20],
-            ]
-        )
-
-        norm = packer._normalize_sheet(
-            raw,
-            helper_names={"index"},
-            reset_index=False,
-            rename_map={"heat_network": "heat_network_lt"},
-        )
-
-        # index column removed, rename applied, index preserved
-        assert list(norm.columns) == ["heat_network_lt", "value"]
-        assert norm.index.tolist() == [2, 3]  # original DataFrame indices kept
-
     def test_safe_get_bool(self):
         """Test _safe_get_bool method"""
         packer = ScenarioPacker()
@@ -781,121 +728,6 @@ class TestScenarioPackerHelpers:
         packer._apply_metadata_to_scenario(scenario, {})
         scenario.update_metadata.assert_not_called()
 
-    def test_extract_scenario_sheet_info_series_and_df(self):
-        """Test _extract_scenario_sheet_info method"""
-        packer = ScenarioPacker()
-
-        ser = pd.Series(
-            {
-                "short_name": "S",
-                "sortables": "SORT1",
-                "custom_curves": "CUR1",
-            },
-            name="COL1",
-        )
-        out = packer._extract_scenario_sheet_info(ser)
-        assert out == {
-            "COL1": {"short_name": "S", "sortables": "SORT1", "custom_curves": "CUR1"}
-        }
-
-        df = pd.DataFrame(
-            {
-                "A": {"short_name": None, "sortables": "S_A", "custom_curves": None},
-                "B": {"short_name": "B_S", "sortables": None, "custom_curves": "C_B"},
-            }
-        )
-        out2 = packer._extract_scenario_sheet_info(df)
-        assert out2["A"]["short_name"] == "A"
-        assert out2["A"]["sortables"] == "S_A"
-        assert out2["A"]["custom_curves"] is None
-        assert out2["B"]["short_name"] == "B_S"
-        assert out2["B"]["custom_curves"] == "C_B"
-
-    def test_process_single_scenario_sortables(self):
-        """Test _process_single_scenario_sortables method"""
-        packer = ScenarioPacker()
-        scenario = Mock(spec=Scenario)
-
-        # Build a sheet where header row contains helpers + target column to be renamed
-        raw = pd.DataFrame(
-            [
-                [None, None, None],
-                ["sortables", "heat_network", "hour"],
-                [None, "lt", 1],
-                [None, "mt", 2],
-            ]
-        )
-
-        packer._process_single_scenario_sortables(scenario, raw)
-        assert scenario.set_sortables_from_dataframe.called
-        df_arg = scenario.set_sortables_from_dataframe.call_args[0][0]
-        assert "heat_network_lt" in df_arg.columns
-        assert "hour" not in df_arg.columns
-
-    def test_process_single_scenario_sortables_empty_after_normalize(self):
-        """Test _process_single_scenario_sortables with empty data after normalization"""
-        packer = ScenarioPacker()
-        scenario = Mock(spec=Scenario)
-
-        raw = pd.DataFrame(
-            [
-                [None, None],
-                ["sortables", "hour"],
-                [None, 1],
-            ]
-        )
-
-        packer._process_single_scenario_sortables(scenario, raw)
-        scenario.set_sortables_from_dataframe.assert_not_called()
-
-    def test_process_single_scenario_curves_success_and_error(self, monkeypatch):
-        """Test _process_single_scenario_curves method"""
-        packer = ScenarioPacker()
-        scenario = Mock(spec=Scenario)
-        scenario.id = 999
-
-        raw = pd.DataFrame(
-            [
-                [None, None],
-                ["custom_curves", "value"],
-                ["curve_1", 1.0],
-                ["curve_2", 2.0],
-            ]
-        )
-
-        dummy_curves = Mock(spec=CustomCurves)
-        monkeypatch.setattr(
-            CustomCurves,
-            "_from_dataframe",
-            staticmethod(lambda df, scenario_id: dummy_curves),
-        )
-        packer._process_single_scenario_curves(scenario, raw)
-        scenario.update_custom_curves.assert_called_once_with(dummy_curves)
-        scenario.update_custom_curves.reset_mock()
-
-        def boom(_df, scenario_id):
-            raise RuntimeError("bad curves")
-
-        monkeypatch.setattr(CustomCurves, "_from_dataframe", staticmethod(boom))
-        packer._process_single_scenario_curves(scenario, raw)
-        scenario.update_custom_curves.assert_not_called()
-
-    def test_process_single_scenario_curves_empty_after_normalize(self):
-        """Test _process_single_scenario_curves with empty data after normalization"""
-        packer = ScenarioPacker()
-        scenario = Mock(spec=Scenario)
-        scenario.id = 1
-
-        raw = pd.DataFrame(
-            [
-                [None],
-                ["custom_curves"],
-                [None],
-            ]
-        )
-        packer._process_single_scenario_curves(scenario, raw)
-        scenario.update_custom_curves.assert_not_called()
-
 
 class TestCreateScenarioFromColumn:
 
@@ -953,328 +785,6 @@ class TestCreateScenarioFromColumn:
         )
         ser = pd.Series({"scenario_id": None, "area_code": None, "end_year": None})
         assert packer._create_scenario_from_column("COL", ser) is None
-
-
-class TestFromExcelDetailed:
-
-    def test_from_excel_full_flow(self, tmp_path, monkeypatch):
-        """Test complete from_excel flow"""
-        # Prepare MAIN with two scenarios: one load, one create
-        main = pd.DataFrame(
-            {
-                "S1": {
-                    "scenario_id": 101,
-                    "area_code": "nl2015",
-                    "end_year": 2050,
-                    "private": "yes",
-                    "template": 3,
-                    "source": "source1",
-                    "title": "Title 1",
-                    "short_name": "Short1",
-                    "sortables": "S1_SORT",
-                    "custom_curves": "S1_CURVES",
-                },
-                "S2": {
-                    "scenario_id": None,
-                    "area_code": "de",
-                    "end_year": 2030,
-                    "private": 0,
-                    "template": None,
-                    "source": "",
-                    "title": "",
-                    "short_name": None,
-                    "sortables": "S2_SORT",
-                    "custom_curves": None,
-                },
-            }
-        )
-
-        # Other sheets
-        params = pd.DataFrame([["helper", "value"], ["input_key", 1]])
-        gqueries = pd.DataFrame([["gquery", "future"], ["co2_emissions", 100]])
-        s1_sort = pd.DataFrame([[None, None], ["sortables", "value"], ["a", 1]])
-        s2_sort = pd.DataFrame([[None, None], ["sortables", "value"], ["b", 2]])
-        s1_curves = pd.DataFrame([[None, None], ["custom_curves", "value"], ["x", 1]])
-
-        path = tmp_path / "import.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-            params.to_excel(
-                writer, sheet_name="SLIDER_SETTINGS", header=False, index=False
-            )
-            gqueries.to_excel(writer, sheet_name="GQUERIES", header=False, index=False)
-            s1_sort.to_excel(writer, sheet_name="S1_SORT", header=False, index=False)
-            s2_sort.to_excel(writer, sheet_name="S2_SORT", header=False, index=False)
-            s1_curves.to_excel(
-                writer, sheet_name="S1_CURVES", header=False, index=False
-            )
-
-        # Patch loading/creating and pack interactions
-        s_loaded = Mock(spec=Scenario)
-        s_loaded.id = "101"
-        s_loaded.identifier = Mock(return_value="101")
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-
-        monkeypatch.setattr(
-            Scenario, "load", staticmethod(lambda *args, **kwargs: s_loaded)
-        )
-        monkeypatch.setattr(
-            Scenario, "new", staticmethod(lambda *args, **kwargs: s_created)
-        )
-
-        # Spy on inputs and queries imports
-        with (
-            patch.object(InputsPack, "set_scenario_short_names") as set_sn,
-            patch.object(InputsPack, "from_dataframe") as from_df,
-            patch.object(QueryPack, "from_dataframe") as gq_from_df,
-            patch.object(
-                ScenarioPacker, "_process_single_scenario_sortables"
-            ) as proc_sort,
-            patch.object(
-                ScenarioPacker, "_process_single_scenario_curves"
-            ) as proc_curves,
-        ):
-            packer = ScenarioPacker.from_excel(str(path))
-
-            assert isinstance(packer, ScenarioPacker)
-            assert s_loaded in packer._scenarios()
-            assert s_created in packer._scenarios()
-
-            set_sn.assert_called_once()
-            from_df.assert_called_once()
-            gq_from_df.assert_called_once()
-
-            # Called once for each scenario with a sortables sheet
-            assert proc_sort.call_count == 2
-            proc_curves.assert_called_once()
-
-    def test_from_excel_missing_or_bad_main(self, tmp_path):
-        """Test from_excel with missing or bad main sheet"""
-        packer = ScenarioPacker.from_excel(str(tmp_path / "bad.xlsx"))
-        assert isinstance(packer, ScenarioPacker)
-        assert len(packer._scenarios()) == 0
-
-        # File with no MAIN sheet
-        path = tmp_path / "no_main.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            pd.DataFrame([[1]]).to_excel(writer, sheet_name="OTHER")
-        packer2 = ScenarioPacker.from_excel(str(path))
-        assert isinstance(packer2, ScenarioPacker)
-        assert len(packer2._scenarios()) == 0
-
-        # File with empty MAIN sheet
-        path2 = tmp_path / "empty_main.xlsx"
-        with pd.ExcelWriter(path2, engine="xlsxwriter") as writer:
-            pd.DataFrame().to_excel(writer, sheet_name="MAIN")
-        packer3 = ScenarioPacker.from_excel(str(path2))
-        assert isinstance(packer3, ScenarioPacker)
-        assert len(packer3._scenarios()) == 0
-
-    def test_from_excel_slider_settings_and_gqueries_errors(
-        self, tmp_path, monkeypatch
-    ):
-        """Test from_excel with errors in slider settings and gqueries import"""
-        main = pd.DataFrame(
-            {
-                "S": {
-                    "scenario_id": None,
-                    "area_code": "nl2015",
-                    "end_year": 2050,
-                    "sortables": None,
-                    "custom_curves": None,
-                }
-            }
-        )
-        params = pd.DataFrame([["helper", "value"], ["input_key", 1]])
-        gqueries = pd.DataFrame([["gquery", "future"], ["co2_emissions", 100]])
-
-        path = tmp_path / "errs.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-            params.to_excel(
-                writer, sheet_name="SLIDER_SETTINGS", header=False, index=False
-            )
-            gqueries.to_excel(writer, sheet_name="GQUERIES", header=False, index=False)
-
-        # Create returns a simple scenario
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-        monkeypatch.setattr(Scenario, "new", staticmethod(lambda a, y: s_created))
-
-        with (
-            patch.object(InputsPack, "set_scenario_short_names") as set_sn,
-            patch.object(
-                InputsPack, "from_dataframe", side_effect=RuntimeError("bad params")
-            ),
-            patch.object(
-                QueryPack, "from_dataframe", side_effect=RuntimeError("bad gq")
-            ),
-        ):
-            packer = ScenarioPacker.from_excel(str(path))
-
-            assert isinstance(packer, ScenarioPacker)
-            # Scenario was still created even if imports failed
-            assert s_created in packer._scenarios()
-            set_sn.assert_called_once()
-
-    def test_from_excel_gqueries_sheet_name_fallback(self, tmp_path, monkeypatch):
-        """Test from_excel with gqueries sheet name fallback"""
-        main = pd.DataFrame(
-            {"S": {"scenario_id": None, "area_code": "nl2015", "end_year": 2050}}
-        )
-
-        path = tmp_path / "gq_fallback.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-            pd.DataFrame([["gquery"], ["total_costs"]]).to_excel(
-                writer, sheet_name="GQ2", header=False, index=False
-            )
-
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-        monkeypatch.setattr(Scenario, "new", staticmethod(lambda a, y: s_created))
-
-        with patch.object(QueryPack, "sheet_name", "GQ2"):
-            with patch.object(QueryPack, "from_dataframe") as gq_from_df:
-                packer = ScenarioPacker.from_excel(str(path))
-                assert s_created in packer._scenarios()
-                gq_from_df.assert_called_once()
-
-    def test_from_excel_processing_sortables_and_curves_errors(
-        self, tmp_path, monkeypatch
-    ):
-        """Test from_excel with errors in processing sortables and curves"""
-        main = pd.DataFrame(
-            {
-                "S": {
-                    "scenario_id": None,
-                    "area_code": "nl2015",
-                    "end_year": 2050,
-                    "sortables": "S_SORT",
-                    "custom_curves": "S_CURVES",
-                }
-            }
-        )
-
-        path = tmp_path / "proc_errs.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-            pd.DataFrame([[None], ["sortables"], ["a"]]).to_excel(
-                writer, sheet_name="S_SORT", header=False, index=False
-            )
-            pd.DataFrame([[None], ["custom_curves"], ["x"]]).to_excel(
-                writer, sheet_name="S_CURVES", header=False, index=False
-            )
-
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-        monkeypatch.setattr(Scenario, "new", staticmethod(lambda a, y: s_created))
-
-        with (
-            patch.object(
-                ScenarioPacker,
-                "_process_single_scenario_sortables",
-                side_effect=RuntimeError("bad sort"),
-            ),
-            patch.object(
-                ScenarioPacker,
-                "_process_single_scenario_curves",
-                side_effect=RuntimeError("bad cur"),
-            ),
-        ):
-            packer = ScenarioPacker.from_excel(str(path))
-            assert isinstance(packer, ScenarioPacker)
-            assert s_created in packer._scenarios()
-
-    def test_from_excel_setup_column_exception_and_all_fail(
-        self, tmp_path, monkeypatch
-    ):
-        """Test from_excel with setup column exceptions"""
-        # Two columns: first raises, second returns scenario
-        main = pd.DataFrame(
-            {
-                "A": {"scenario_id": None, "area_code": "nl2015", "end_year": 2050},
-                "B": {"scenario_id": None, "area_code": "de", "end_year": 2030},
-            }
-        )
-        path = tmp_path / "columns_mix.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-
-        # Patch method to raise for A and create for B
-        def setup(col_name, col_ser):
-            if col_name == "A":
-                raise RuntimeError("boom")
-            s = Mock(spec=Scenario)
-            s.id = "BID"
-            s.identifier = Mock(return_value="BID")
-            return s
-
-        with patch.object(
-            ScenarioPacker, "_create_scenario_from_column", side_effect=setup
-        ):
-            packer = ScenarioPacker.from_excel(str(path))
-            assert any(s.id == "BID" for s in packer._scenarios())
-
-        # All columns fail -> 0 scenarios, early return
-        with patch.object(
-            ScenarioPacker,
-            "_create_scenario_from_column",
-            side_effect=RuntimeError("e"),
-        ):
-            packer2 = ScenarioPacker.from_excel(str(path))
-            assert len(packer2._scenarios()) == 0
-
-    def test_from_excel_missing_slider_settings_sheet_parse_error(
-        self, tmp_path, monkeypatch
-    ):
-        """Test from_excel with missing slider settings sheet"""
-        main = pd.DataFrame(
-            {"S": {"scenario_id": None, "area_code": "nl2015", "end_year": 2050}}
-        )
-        path = tmp_path / "no_params.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-        monkeypatch.setattr(Scenario, "new", staticmethod(lambda a, y: s_created))
-        packer = ScenarioPacker.from_excel(str(path))
-        assert s_created in packer._scenarios()
-
-    def test_from_excel_gqueries_parse_raises(self, tmp_path, monkeypatch):
-        """Test from_excel with gqueries parse error"""
-        main = pd.DataFrame(
-            {"S": {"scenario_id": None, "area_code": "nl2015", "end_year": 2050}}
-        )
-        path = tmp_path / "gq_parse_err.xlsx"
-        with pd.ExcelWriter(path, engine="xlsxwriter") as writer:
-            main.to_excel(writer, sheet_name="MAIN")
-            pd.DataFrame([["gquery"], ["total_costs"]]).to_excel(
-                writer, sheet_name="GQUERIES", header=False, index=False
-            )
-
-        s_created = Mock(spec=Scenario)
-        s_created.id = "created"
-        s_created.identifier = Mock(return_value="created")
-        monkeypatch.setattr(Scenario, "new", staticmethod(lambda a, y: s_created))
-
-        original_parse = pd.ExcelFile.parse
-
-        def parse_proxy(self, sheet_name, *a, **k):
-            if sheet_name == "GQUERIES":
-                raise ValueError("bad parse")
-            return original_parse(self, sheet_name, *a, **k)
-
-        with patch.object(pd.ExcelFile, "parse", parse_proxy):
-            packer = ScenarioPacker.from_excel(str(path))
-            assert s_created in packer._scenarios()
 
 
 class TestInputsPackIntegration:
@@ -1387,7 +897,7 @@ class TestScenarioPackerExtras:
                 "to_dataframe",
                 return_value=pd.DataFrame({"v": [1]}),
             ),
-            patch("pyetm.models.scenario_packer.add_frame") as add_frame,
+            patch("pyetm.utils.excel_utils.add_frame") as add_frame,
             patch("pyetm.models.scenario_packer.Workbook") as mock_wb,
         ):
             mock_wb.return_value = Mock()
@@ -1446,104 +956,6 @@ class TestScenarioPackerExtras:
             args2, _ = toe2.call_args
             assert args2[1] == ["h2"]
 
-    def test_build_excel_main_dataframe_applies_labels_and_order(self):
-        packer = ScenarioPacker()
-        s1 = Mock(spec=Scenario)
-        s1.id = "1"
-        s1.identifier = Mock(return_value="Label 1")
-        s2 = Mock(spec=Scenario)
-        s2.id = "2"
-        s2.identifier = Mock(return_value="Label 2")
-        packer.add(s1, s2)
-
-        df = pd.DataFrame(
-            {"1": ["A", 2050], "2": ["B", 2040]}, index=["area_code", "end_year"]
-        )
-
-        with patch.object(ScenarioPacker, "main_info", return_value=df):
-            out = packer._build_excel_main_dataframe()
-            # Columns should be relabeled using identifier
-            assert list(out.columns) == ["Label 1", "Label 2"]
-            # Preferred ordering keeps known fields order among those present
-            assert out.index.tolist()[0:2] == ["area_code", "end_year"]
-
-    def test_sanitize_dataframe_for_excel_handles_datetime_and_objects(self):
-        packer = ScenarioPacker()
-        import datetime as dt
-
-        class Foo:
-            def __str__(self):
-                return "foo"
-
-        frame = pd.DataFrame(
-            {
-                "col": [dt.datetime(2020, 1, 1), Foo()],
-            },
-            index=["when", "obj"],
-        )
-        out = packer._sanitize_dataframe_for_excel(frame)
-        assert out.loc["obj", "col"] == "foo"
-
-    def test_log_scenario_warnings_helper(self):
-        packer = ScenarioPacker()
-        s = Mock(spec=Scenario)
-        s.identifier = Mock(return_value="SID")
-        helper_attr = Mock()
-        s._sortables = helper_attr
-        packer._log_scenario_warnings(s, "_sortables", "Sortables")
-        helper_attr.log_warnings.assert_called_once()
-
-
-def test_apply_scenario_column_labels_no_matches():
-    packer = ScenarioPacker()
-    # No scenarios added; rename map should be empty and df unchanged
-    df = pd.DataFrame({"X": [1]}, index=["row"])
-    out = packer._apply_scenario_column_labels(df)
-    assert list(out.columns) == ["X"]
-
-
-def test_build_column_rename_map_both_match_str_and_exact():
-    packer = ScenarioPacker()
-    s1 = Mock(spec=Scenario)
-    s1.id = "1"
-    s2 = Mock(spec=Scenario)
-    s2.id = 2  # int id for exact match case
-    # Columns include string '1' and int 2
-    cols = ["1", 2, "nope"]
-    rename = packer._build_column_rename_map([s1, s2], cols)
-    assert "1" in rename and 2 in rename
-    # 'nope' should not be in rename map
-    assert "nope" not in rename
-
-
-def test_get_scenario_display_label_fallbacks():
-    packer = ScenarioPacker()
-    # 1) identifier raises -> title used
-    sc = Mock(spec=Scenario)
-    sc.identifier = Mock(side_effect=RuntimeError("bad"))
-    sc.title = "T"
-    assert packer._get_scenario_display_label(sc, "FALL") == "T"
-
-    # 2) no title -> id used
-    sc2 = Mock(spec=Scenario)
-    sc2.identifier = Mock(side_effect=RuntimeError("bad"))
-    sc2.title = None
-    sc2.id = 42
-    assert packer._get_scenario_display_label(sc2, "FALL") == "42"
-
-    # 3) no id -> fallback to column
-    sc3 = Mock(spec=Scenario)
-    sc3.identifier = Mock(side_effect=RuntimeError("bad"))
-    sc3.title = None
-    sc3.id = None
-    assert packer._get_scenario_display_label(sc3, "COLX") == "COLX"
-
-
-def test_get_value_before_output_respects_output_boundary():
-    packer = ScenarioPacker()
-    series = pd.Series(["S1", None, "S2"], index=["sortables", "output", "sortables"])
-    assert packer._get_value_before_output(series, "sortables") == "S1"
-
 
 def test_export_output_curves_if_needed_false():
     packer = ScenarioPacker()
@@ -1565,7 +977,7 @@ def test_export_output_curves_if_needed_false():
 def test_add_gqueries_sheet_disabled():
     packer = ScenarioPacker()
     with (
-        patch("pyetm.models.scenario_packer.add_frame") as add_frame,
+        patch("pyetm.utils.excel_utils.add_frame") as add_frame,
         patch("pyetm.models.scenario_packer.Workbook") as wb,
         patch.object(ScenarioPacker, "_scenarios", return_value={Mock(spec=Scenario)}),
         patch.object(ScenarioPacker, "_add_data_sheets"),
